+++ src/Reducer/index.js
@@ -1,0 +1,3 @@
+import { filterDataReducer } from "./filterData.reducer";
+
+export {filterDataReducer};
